Class 'FavoriteHolidaysEditor' moved to package editor.
