+++ src/main/java/com/infoshareacademy/App.java
@@ -1,11 +1,7 @@
 package com.infoshareacademy;
 
-import com.infoshareacademy.api.FavoriteHolidaysEditor;
-import com.infoshareacademy.api.Printer;
-import com.infoshareacademy.edit.HolidaysEditor;
+import com.infoshareacademy.edit.FavoriteHolidaysEditor;
 import com.infoshareacademy.api.HolidaysJsonData;
-import com.infoshareacademy.menu.MainMenu;
-import com.infoshareacademy.menu.MenuSearch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
