+++ src/main/java/com/infoshareacademy/edit/FavoriteHolidaysEditor.java
@@ -1,4 +1,7 @@
-package com.infoshareacademy.api;
+package com.infoshareacademy.edit;
+
+import com.infoshareacademy.api.Holidays;
+import com.infoshareacademy.api.HolidaysJsonData;
 
 import java.util.*;
